+++ src/core/hook/use-concent.js
@@ -97,7 +97,9 @@ export default function useConcent(registerOption, ccClassKey){
   }
   
   const refCtx = hookRef.ctx;
+  refCtx.prevProps = refCtx.props;
   refCtx.props = props;
+  hookRef.props = props;
 
   // ???does user really need beforeMount,mounted,beforeUpdate,updated,beforeUnmount in setup???